[MOS-755] Wrong msleep duration fix
Fix of the issue that msleep() function
waited twice the requested time due to
discrepancy between real PIT clock source
and the one used to configure the timer.
Fixing that caused another issue - PIT
IRQ handler was called twice every timer
overflow due to not waiting for status
flag to be cleared.