[MOS-806] Fixed issue with lseek and SEEK_END
lwext4 does not support negative offset values.In case of using SEEK_END, internally, it always subtracts passed value from the current file position. This is valid behavior as it won't allow for setting file pointer past the end of file, but at the same time, is not coherent with C stdlib.